1
прибрав відкриття модалки з сайдбару та посилання у тексті
